Refactor(trending): modularize Firestore utilities and clarify responsibilities
